staging: unisys: Remove unneeded fields in diagchannel.h
Diagchannel.h is used primarily for the diagnostics channel. The
diagnostics channel is not being used by linux guests currently, so
the majority of the file is not needed. What is left is what is
needed to perform postcode vmcalls. Those postcodes will eventually
end up in the diag channel.

Signed-off-by: David Kershner <david.kershner@unisys.com>
Signed-off-by: Benjamin Romer <benjamin.romer@unisys.com>

Conflicts:
	drivers/staging/unisys/visorbus/visorbus_main.c
	drivers/staging/unisys/visorbus/visorchipset.c