staging: unisys: make module unremoveable
Once a netdevice is registered, theres nothing to unregister it, even if
the module is removed. That's obviously a memory leak and an easy oops
vector.

Fixing it isn't hard, but it requires that some additional infrastructure
be put in place to handle device removal and shutdown that isn't in the
driver yet. For now we simply remove the module cleanup routine. This has
the effect of preventing the module from being removed, so we never hit
the problem.

Signed-off-by: Neil Horman <nhorman@redhat.com>
Signed-off-by: Benjamin Romer <benjamin.romer@unisys.com>