staging: unisys: Capture data from device create to register interrupt.
s-Par Firmware sends a CREATE_DEVICE information for each device that
gets created. Contained in the CREATE_DEVICE is the GSI interrupt vector
we should request an interrupt on. Save off that information and when
the client driver asks to register for interrupt, register the gsi
number and then request the irq based off that information.

When an interrupt happens, update the flag to inform IO Service Partition
and call the channel_interrupt function. Update the flag in the rearm
function to accept another interrupt.