staging: unisys: dont dynamically allocate in data path
Visorhba was allocating and deallocating cmdrsps in the datapath. This
should not be allowed. Instead preallocate a given set of cmdrsps for
the device. The scsipending array was used to store the preallocated
cmdrsps because it maintains the list of maximum commands that can be
in the Service Partition at a given time.

Modified the behavior of add_scsipending_entry to accept a NULL value
instead of the cmdrsp pointer. If a NULL is passed in via the new field
the existing sent field will just point to the cmdrsp in the pending
entry.

Created function get_scsipending_cmdrsp that retrieves the cmdrsp
from the pending_array given the device data and the entry number.

Conflicts:
	drivers/staging/unisys/visorhba/visorhba_main.c