staging: unisys: visorinput: make structs & arrays const where possible
This also gave me a warning with the assignment:

    visorinput_dev->keycode = visorkbd_keycode;

because input_dev->keycode is NOT static but visorkbd_keycode now is, so
I went ahead and also added logic to stash away non-static copies of
visorkbd_keycode[] and visorkbd_ext_keycode[] within visorinput_devdata,
and use the copy to assign to visorinput_dev->keycode.  This change is also
technically required, because user-space can remap keys, and we don't want
this to be shared with the other keyboard devices running on the same system.