staging: unisys: visorinput: remove need for 'depends on FB'
Previously, we used a hack to determine the max x,y resolution of the visor
virtual mouse: we just looked at the resolution of the first-registered
framebuffer device, using the currently-valid assumption that in a Unisys
s-Par guest environment the video will be provided by an efifb framebuffer
device.

This hack has been removed, by instead determining the default mouse
resolution by looking at fields within the visor mouse channel memory,
mouse.x_resolution and mouse.y_resolution.  If these fields are 0, a default
resolution of 1024x768 is assumed.  Note that in current implementations,
mouse.x_resolution and mouse.y_resolution are always just initialized to 0 by
the back-end, and only 1024x768 is supported.  But coding it this way will
allow other resolutions to work in the future.