staging: unisys: fix visorhba_probe() deadlock in scsi_scan_host()
This patch needs to be split up into 3 or 4 little patches.  I am committing
this as a preview version, just because it works.

For some reason, scsi_scan_host() now needs to do I/Os when it is called by
visorhba_probe().  It's hard for me to believe that something else hasn't gone
into the kernel between 4.3.0-rc4 and 4.3.0-rc7 to make this happen, because
I can't see how visorhba could have ever handled this situation before, even
prior to David's interrupt code.

Without this patch, this scenario causes a deadlock, because visorhba_probe()
grabs visordriver_callback_lock(), and we attempt to go after this same lock
before calling the function driver's channel_interrupt() function.  That means
you can NOT complete an I/O within visorhba_probe(), and if you initiate an I/O
and wait for it to complete, you're going to be waiting forever.

The new interrupt code was NOT distinguishing between the availability of an
irq (i.e., visor_device.irq != 0), and the fact that we were in fact operating
in real interrupt mode (i.e., request_irq() succeeded).  This could cause us
to do the wrong thing in visorbus_enable_channel_interrupts(),
visorbus_disable_channel_interrupts(), and visorbus_rearm_channel_interrupts().
This was fixed by adding a boolean named visor_device.irq_requested, which is
true iff request_irq() succeeded.

We were not properly restoring interrupt-related state after a failed attempt
of visorbus_register_channel_interrupts(), nor after the device was
unattached from a function driver.  Most-notably, we did NOT call free_irq()
(the book-end to request_irq()).  This was fixed via a new
visorbus_unregister_for_channel_interrupts().

We can't complete I/Os between calling
visorbus_register_for_channel_interrupts() and setting the channel state to
CHANNELCLI_OWNED, because they will never complete.  The reason they will never
complete is because:
* visorbus is NOT polling because it thinks real interrupts are enabled
* the IO partition will NOT send interrupts because the channel state is
  not yet CHANNELCLI_OWNED
But this is exactly the situation as it existed at the end of visorhba_probe(),
given that the scsi_scan_host() can do I/Os.  So I hacked the end of
visorhba_probe() so that everyone essentially agrees that we can complete I/Os
in polling mode within scsi_scan_host().

TBD: I don't think the special-case I added to the end of visorhba_probe()
is the right way to go.  I think that it should be visorbus that in fact
calls visorbus_register_for_channel_interrupts(), after a successful return
of the function driver probe(), but before setting the channel state to
CHANNELCLI_OWNED.  That way, we don't have to resort to 'slippery-coding' in
every function driver probe() function that wants to do I/Os.

Also, error messages were added to reveal other possible error paths in
visorbus.